MNT: move set_many to utils
